AddNewFunctions

+++ database/migrations/2021_02_23_194604_create_students_table.php
@@ -25,7 +25,7 @@ class CreateStudentsTable extends Migration
              $table->integer('year')->nullable();
              $table->integer('number')->nullable();
              $table->string('email');
-             $table->integer('mobile')->nullable();
+             $table->bigInteger('mobile')->nullable();
              $table->integer('password')->nullable();
             $table->timestamps();
         });
